better if

+++ src/bkg.sh
@@ -62,8 +62,11 @@ main() {
         set_BKG BKG_MIN_CALLS_TO_API "0"
     fi
 
-    [ ! -f "$BKG_INDEX_DB" ] || mv "$BKG_INDEX_DB" "$BKG_INDEX_DB".bak
-    [ ! -f "$BKG_INDEX_SQL.zst" ] || unzstd -v -c "$BKG_INDEX_SQL.zst" | sqlite3 "$BKG_INDEX_DB"
+    if [ -f "$BKG_INDEX_SQL.zst" ]; then
+        [ ! -f "$BKG_INDEX_DB" ] || mv "$BKG_INDEX_DB" "$BKG_INDEX_DB".bak
+        unzstd -c "$BKG_INDEX_SQL.zst" | sqlite3 "$BKG_INDEX_DB"
+    fi
+
     [ -f "$BKG_INDEX_DB" ] || {
         [ -f "$BKG_INDEX_DB".bak ] && mv "$BKG_INDEX_DB".bak "$BKG_INDEX_DB" || sqlite3 "$BKG_INDEX_DB" ""
     }
